feat(ender): some refactoring
