+++ ee/backend/cmd/ender/main.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"os"
 	"os/signal"
-	"strings"
 	"syscall"
 	"time"
 
@@ -24,7 +23,6 @@ import (
 	"openreplay/backend/pkg/queue"
 	"openreplay/backend/pkg/queue/types"
 	"openreplay/backend/pkg/sessions"
-	"openreplay/backend/pkg/storage"
 )
 
 func main() {
@@ -34,7 +32,6 @@ func main() {
 
 	h := health.New()
 
-	// Observability
 	dbMetric := database.New("ender")
 	enderMetric := enderMetrics.New("ender")
 	metrics.New(log, append(enderMetric.List(), dbMetric.List()...))
@@ -125,110 +122,9 @@ func main() {
 			cleanupDispatcher.Close()
 			os.Exit(0)
 		case <-tick:
-			details := newDetails()
-
-			// Find ended sessions and send notification to other services
-			sessionEndGenerator.HandleEndedSessions(func(sessions map[uint64]uint64) map[uint64]bool {
-				// Load all sessions from DB
-				sessionsList := make([]uint64, 0, len(sessions))
-				for sessionID := range sessions {
-					sessionsList = append(sessionsList, sessionID)
-				}
-				completedSessions := make(map[uint64]bool)
-				sessionsData, err := sessManager.GetManySessions(sessionsList)
-				if err != nil {
-					log.Error(ctx, "can't get sessions from database: %s", err)
-					return completedSessions
-				}
-
-				// Check if each session was ended
-				for sessionID, sess := range sessionsData {
-					sessCtx := context.WithValue(context.Background(), "sessionID", fmt.Sprintf("%d", sessionID))
-
-					timestamp := sessions[sessionID]
-					var currDuration uint64 = 0
-					if sess.Duration != nil {
-						currDuration = *sess.Duration
-					}
-					newDur := timestamp - sess.Timestamp
-
-					// Skip if session was ended before with same duration
-					if currDuration == newDur {
-						details.Duplicated[sessionID] = currDuration
-						completedSessions[sessionID] = true
-						continue
-					}
-					if currDuration > newDur {
-						details.Shorter[sessionID] = int64(currDuration) - int64(newDur)
-						completedSessions[sessionID] = true
-						continue
-					}
-
-					newDuration, err := sessManager.UpdateDuration(sessionID, timestamp)
-					if err != nil {
-						if strings.Contains(err.Error(), "integer out of range") {
-							// Skip session with broken duration
-							details.Failed[sessionID] = timestamp
-							completedSessions[sessionID] = true
-							continue
-						}
-						if strings.Contains(err.Error(), "is less than zero for uint64") {
-							details.Negative[sessionID] = timestamp
-							completedSessions[sessionID] = true
-							continue
-						}
-						if strings.Contains(err.Error(), "no rows in result set") {
-							details.NotFound[sessionID] = timestamp
-							completedSessions[sessionID] = true
-							continue
-						}
-						log.Error(sessCtx, "can't update session duration, err: %s", err)
-						continue
-					}
-					// Check one more time just in case
-					if currDuration == newDuration {
-						details.Duplicated[sessionID] = currDuration
-						completedSessions[sessionID] = true
-						continue
-					}
-					msg := &messages.SessionEnd{Timestamp: timestamp}
-					if cfg.UseEncryption {
-						if key := storage.GenerateEncryptionKey(); key != nil {
-							if err := sessManager.UpdateEncryptionKey(sessionID, key); err != nil {
-								log.Warn(sessCtx, "can't save session encryption key: %s, session will not be encrypted", err)
-							} else {
-								msg.EncryptionKey = string(key)
-							}
-						}
-					}
-					if sess != nil && (sess.Platform == "ios" || sess.Platform == "android") {
-						msg := &messages.MobileSessionEnd{Timestamp: timestamp}
-						if err := producer.Produce(cfg.TopicRawMobile, sessionID, msg.Encode()); err != nil {
-							log.Error(sessCtx, "can't send MobileSessionEnd to mobile topic: %s", err)
-							continue
-						}
-						if err := producer.Produce(cfg.TopicRawImages, sessionID, msg.Encode()); err != nil {
-							log.Error(sessCtx, "can't send MobileSessionEnd signal to canvas topic: %s", err)
-						}
-					} else {
-						if err := producer.Produce(cfg.TopicRawAssets, sessionID, msg.Encode()); err != nil {
-							log.Error(sessCtx, "can't send sessionEnd to raw topic: %s", err)
-							continue
-						}
-						if err := producer.Produce(cfg.TopicCanvasImages, sessionID, msg.Encode()); err != nil {
-							log.Error(sessCtx, "can't send sessionEnd signal to canvas topic: %s", err)
-						}
-					}
-
-					if currDuration != 0 {
-						details.Diff[sessionID] = int64(newDuration) - int64(currDuration)
-						details.Updated++
-					} else {
-						details.New++
-					}
-					completedSessions[sessionID] = true
-				}
-				return completedSessions
+			details := ender.NewLogDetails()
+			sessionEndGenerator.HandleEndedSessions(func(candidates map[uint64]uint64) map[uint64]bool {
+				return processEndedBatch(ctx, candidates, sessManager, producer, cfg, log, details)
 			})
 			details.Log(log, ctx)
 			producer.Flush(cfg.ProducerTimeout)
@@ -243,86 +139,34 @@ func main() {
 	}
 }
 
-type logDetails struct {
-	Failed     map[uint64]uint64
-	Duplicated map[uint64]uint64
-	Negative   map[uint64]uint64
-	Shorter    map[uint64]int64
-	NotFound   map[uint64]uint64
-	Diff       map[uint64]int64
-	Updated    int
-	New        int
+func processEndedBatch(
+	ctx context.Context,
+	candidates map[uint64]uint64,
+	sessManager sessions.Sessions,
+	producer types.Producer,
+	cfg *config.Config,
+	log logger.Logger,
+	details *ender.LogDetails,
+) map[uint64]bool {
+	completed := make(map[uint64]bool, len(candidates))
+	if len(candidates) == 0 {
+		return completed
+	}
+	ids := make([]uint64, 0, len(candidates))
+	for sessionID := range candidates {
+		ids = append(ids, sessionID)
+	}
+	loaded, err := sessManager.GetManySessions(ids)
+	if err != nil {
+		log.Error(ctx, "can't get sessions from database: %s", err)
+		return completed
+	}
+	for sessionID, timestamp := range candidates {
+		sessCtx := context.WithValue(ctx, "sessionID", fmt.Sprintf("%d", sessionID))
+		sess := loaded[sessionID]
+		if ender.ProcessEndedSession(sessCtx, sessionID, timestamp, sess, sessManager, producer, cfg, log, details) {
+			completed[sessionID] = true
+		}
+	}
+	return completed
 }
-
-func newDetails() *logDetails {
-	return &logDetails{
-		Failed:     make(map[uint64]uint64),
-		Duplicated: make(map[uint64]uint64),
-		Negative:   make(map[uint64]uint64),
-		Shorter:    make(map[uint64]int64),
-		NotFound:   make(map[uint64]uint64),
-		Diff:       make(map[uint64]int64),
-		Updated:    0,
-		New:        0,
-	}
-}
-
-func (l *logDetails) Log(log logger.Logger, ctx context.Context) {
-	if n := len(l.Failed); n > 0 {
-		log.Debug(ctx, "sessions with wrong duration: %d, %v", n, l.Failed)
-	}
-	if n := len(l.Negative); n > 0 {
-		log.Debug(ctx, "sessions with negative duration: %d, %v", n, l.Negative)
-	}
-	if n := len(l.NotFound); n > 0 {
-		log.Debug(ctx, "sessions without info in DB: %d, %v", n, l.NotFound)
-	}
-	var logBuilder strings.Builder
-	logValues := []interface{}{}
-
-	if len(l.Failed) > 0 {
-		logBuilder.WriteString("failed: %d, ")
-		logValues = append(logValues, len(l.Failed))
-	}
-	if len(l.Negative) > 0 {
-		logBuilder.WriteString("negative: %d, ")
-		logValues = append(logValues, len(l.Negative))
-	}
-	if len(l.Shorter) > 0 {
-		logBuilder.WriteString("shorter: %d, ")
-		logValues = append(logValues, len(l.Shorter))
-	}
-	if len(l.Duplicated) > 0 {
-		logBuilder.WriteString("same: %d, ")
-		logValues = append(logValues, len(l.Duplicated))
-	}
-	if l.Updated > 0 {
-		logBuilder.WriteString("updated: %d, ")
-		logValues = append(logValues, l.Updated)
-	}
-	if l.New > 0 {
-		logBuilder.WriteString("new: %d, ")
-		logValues = append(logValues, l.New)
-	}
-	if len(l.NotFound) > 0 {
-		logBuilder.WriteString("not found: %d, ")
-		logValues = append(logValues, len(l.NotFound))
-	}
-
-	if logBuilder.Len() > 0 {
-		logMessage := logBuilder.String()
-		logMessage = logMessage[:len(logMessage)-2]
-		log.Info(ctx, logMessage, logValues...)
-	}
-}
-
-type SessionEndType int
-
-const (
-	FailedSessionEnd SessionEndType = iota + 1
-	DuplicatedSessionEnd
-	NegativeDuration
-	ShorterDuration
-	NewSessionEnd
-	NoSessionInDB
-)
